PV3.Condition.ConvertToSBV

Documentation

errorNotDefined :: [Char]

quant

Arguments

:: (String -> Symbolic (SBV Integer))
Type of quantification that will be applied (universal/existential).

-> String
Name of variable that will be quantified.

-> (Map String SBool -> Map String SInteger -> Symbolic SBool)
Boolean subexpression, that has already been converted to SBV format. The updated environments will be passed down to this subexpression.

-> Map String SBool
Environment for looking up boolean variables. Will be passed down to the given subexpression.

-> Map String SInteger
Environment for looking up int variables. Will be updated with the variable that will be quantified. Will be passed down to the given subexpression.

-> Symbolic SBool
Resulting boolean subexpression (given subexpression quantified).

Quantifies the given variable (name), bounds it, inserts it into the (given) environment and passes down the (updated) environments to the (already to SBV converted) boolean subexpression. Returns the resulting boolean subexpression.

lookupEnv

Arguments

:: String
Name of the variable that will be looked up.

-> Map String v
Environment in which the given variables will be looked up.

-> v
Looked up variable.

Looks up the given variable in the given environment. Gives an error when the variable is not defined in the given environment.

sem_Cond :: Cond -> T_Cond

type T_Cond = Map String SBool -> Map String SInteger -> Symbolic SBool

data Inh_Cond

Constructors

Inh_Cond

data Syn_Cond

Constructors

Syn_Cond

Fields

symCond_Syn_Cond :: Map String SBool -> Map String SInteger -> Symbolic SBool

wrap_Cond :: T_Cond -> Inh_Cond -> Syn_Cond

sem_Cond_CAnd :: T_Cond -> T_Cond -> T_Cond

sem_Cond_COr :: T_Cond -> T_Cond -> T_Cond

sem_Cond_CNot :: T_Cond -> T_Cond

sem_Cond_CImplies :: T_Cond -> T_Cond -> T_Cond

sem_Cond_CEquiv :: T_Cond -> T_Cond -> T_Cond

sem_Cond_CEQ :: T_IExpr -> T_IExpr -> T_Cond

sem_Cond_CNEQ :: T_IExpr -> T_IExpr -> T_Cond

sem_Cond_CLT :: T_IExpr -> T_IExpr -> T_Cond

sem_Cond_CLTE :: T_IExpr -> T_IExpr -> T_Cond

sem_Cond_CGT :: T_IExpr -> T_IExpr -> T_Cond

sem_Cond_CGTE :: T_IExpr -> T_IExpr -> T_Cond

sem_Cond_CUnivQuant :: String -> T_Cond -> T_Cond

sem_Cond_CExistQuant :: String -> T_Cond -> T_Cond

sem_Cond_CParamOld :: Integer -> T_Cond

sem_Cond_CReturnValue :: T_Cond

sem_Cond_CLit :: Bool -> T_Cond

sem_Cond_CLocal :: Integer -> T_Cond

sem_Cond_CParamCurr :: Integer -> T_Cond

sem_Cond_CStack :: T_IExpr -> T_Cond

sem_IExpr :: IExpr -> T_IExpr

type T_IExpr = Map String SBool -> Map String SInteger -> Symbolic SInteger

data Inh_IExpr

Constructors

Inh_IExpr

data Syn_IExpr

Constructors

Syn_IExpr

Fields

symIExpr_Syn_IExpr :: Map String SBool -> Map String SInteger -> Symbolic SInteger

wrap_IExpr :: T_IExpr -> Inh_IExpr -> Syn_IExpr

sem_IExpr_IAdd :: T_IExpr -> T_IExpr -> T_IExpr

sem_IExpr_ISub :: T_IExpr -> T_IExpr -> T_IExpr

sem_IExpr_IMul :: T_IExpr -> T_IExpr -> T_IExpr

sem_IExpr_IDiv :: T_IExpr -> T_IExpr -> T_IExpr

sem_IExpr_IParamOld :: Integer -> T_IExpr

sem_IExpr_IReturnValue :: T_IExpr

sem_IExpr_ILit :: Integer -> T_IExpr

sem_IExpr_IVarName :: String -> T_IExpr

sem_IExpr_ILocal :: Integer -> T_IExpr

sem_IExpr_IParamCurr :: Integer -> T_IExpr

sem_IExpr_IStack :: T_IExpr -> T_IExpr

sem_IExpr_ITopStackIndex :: Integer -> T_IExpr