PV3.Examples.AnotherExample

Description

Example that uses most instructions and the control structures. The program multiplies its second argument by two.

int P(a0,a1) { var x0 = 10; var x1 = 10; if (10 == 10) x0 = 10; else x0 = 11; a1 = a0 + 10; if (a0 >= 10) {} else return 0; return a1 * 2; }

NB: there are a few instruction here which modify the stack, for which there is no Java-like equivalent

Pre: true Post: (a_0 >= 10 ==> return == (a_0 + 10) * 2) && (a_0 10 == return == 0)

Part of PV - Project 3.

Authors: Jaap van der Plas and Danny Bergsma Version: 0.1, 3 April 2013

Documentation

program :: Program

postcondition :: Cond

The postcondition is:

precondition :: Cond