PV3.WP

Documentation

replaceParamWithOld

Arguments

:: Integer
Number of parameters.

-> Cond
Condition in which occurrences of param_i (0 <= i < #params) will be replaced with a_i.

-> Cond
Resulting condition, which all replacements done.

Replaces, in the given condition (boolean expression), all param_0, param_1, ..., param_(given params - 1). (param_i References to (the (current) value of) param i. At the begin of the program these values are the same as the pre-invocation ones, which can be referenced in the specified pre- and postcondition and thus in the resulting weakest precondition and verification condition. To make sure that param_i and a_i reference to the same value, we rename the first.)

replace

Arguments

:: Cond
Condition in which the given replacements will be done.

-> Maybe (Cond, Cond)
If Just: all occurrences of the first element in given condition will be replaced with the second element. If Nothing: no boolean expressions will be replaced.

-> Maybe (IExpr, IExpr)
If Just: all occurrences of the first element in given condition will be replaced with the second element. If Nothing: no int expressions will be replaced.

-> Cond
Resulting condition, with all replacements done.

binaryOpI

Arguments

:: Cond
The (intermediate) condition the wp rule will be applied on.

-> (IExpr -> IExpr -> IExpr)
The binary operator (result type: int).

-> Cond
The resulting condition, with the wp rule applied on.

Implements the wp rule for given binary operator that has int as result type (+, -, ...).

binaryOpB

Arguments

:: Cond
The (intermediate) condition the wp rule will be applied on.

-> (IExpr -> IExpr -> Cond)
The binary operator (result type: bool).

-> Cond
The resulting condition, with the wp rule applied on.

Implements the wp rule for given binary operator that has bool as result type (>, >=, ...).

sem_Instruction :: Instruction -> T_Instruction

type T_Instruction = Cond -> Cond -> (Instruction, Int, Cond)

data Inh_Instruction

Constructors

Inh_Instruction

Fields

wp_Inh_Instruction :: Cond

wp'_Inh_Instruction :: Cond

data Syn_Instruction

Constructors

Syn_Instruction

Fields

copy_Syn_Instruction :: Instruction

length_Syn_Instruction :: Int

wp_Syn_Instruction :: Cond

wrap_Instruction :: T_Instruction -> Inh_Instruction -> Syn_Instruction

sem_Instruction_InstSetLocal :: Integer -> T_Literal -> T_Instruction

sem_Instruction_InstLoadLocal :: Integer -> T_Instruction

sem_Instruction_InstStoreLocal :: Integer -> T_Instruction

sem_Instruction_InstLoadParam :: Integer -> T_Instruction

sem_Instruction_InstStoreParam :: Integer -> T_Instruction

sem_Instruction_InstPushLiteral :: T_Literal -> T_Instruction

sem_Instruction_InstPop :: T_Instruction

sem_Instruction_InstAdd :: T_Instruction

sem_Instruction_InstMin :: T_Instruction

sem_Instruction_InstMul :: T_Instruction

sem_Instruction_InstLT :: T_Instruction

sem_Instruction_InstLTE :: T_Instruction

sem_Instruction_InstGT :: T_Instruction

sem_Instruction_InstGTE :: T_Instruction

sem_Instruction_InstEQ :: T_Instruction

sem_Instruction_InstEquiv :: T_Instruction

sem_Instruction_Instreturn :: T_Instruction

sem_Literal :: Literal -> T_Literal

type T_Literal = Literal

data Inh_Literal

Constructors

Inh_Literal

data Syn_Literal

Constructors

Syn_Literal

Fields

copy_Syn_Literal :: Literal

wrap_Literal :: T_Literal -> Inh_Literal -> Syn_Literal

sem_Literal_LBool :: Bool -> T_Literal

sem_Literal_LInt :: Integer -> T_Literal

sem_Program :: Program -> T_Program

type T_Program = [Int] -> Cond -> (Int, [Int], Cond)

data Inh_Program

Constructors

Inh_Program

Fields

loopIterations_Inh_Program :: [Int]

wp_Inh_Program :: Cond

data Syn_Program

Constructors

Syn_Program

Fields

length_Syn_Program :: Int

loopIterations_Syn_Program :: [Int]

wp_Syn_Program :: Cond

wrap_Program :: T_Program -> Inh_Program -> Syn_Program

sem_Program_Program :: Integer -> Integer -> T_StatementList -> T_Program

sem_Statement :: Statement -> T_Statement

type T_Statement = [Int] -> Cond -> Cond -> (Statement, Int, [Int], Cond)

data Inh_Statement

Constructors

Inh_Statement

Fields

loopIterations_Inh_Statement :: [Int]

wp_Inh_Statement :: Cond

wp'_Inh_Statement :: Cond

data Syn_Statement

Constructors

Syn_Statement

Fields

copy_Syn_Statement :: Statement

length_Syn_Statement :: Int

loopIterations_Syn_Statement :: [Int]

wp_Syn_Statement :: Cond

wrap_Statement :: T_Statement -> Inh_Statement -> Syn_Statement

sem_Statement_SInstruction :: T_Instruction -> T_Statement

sem_Statement_SIfTrueElse :: T_StatementList -> T_StatementList -> T_Statement

sem_Statement_SWhileTrue :: T_StatementList -> T_Statement

sem_StatementList :: StatementList -> T_StatementList

type T_StatementList = [Int] -> Cond -> Cond -> (StatementList, Int, [Int], Cond)

data Inh_StatementList

Constructors

Inh_StatementList

Fields

loopIterations_Inh_StatementList :: [Int]

wp_Inh_StatementList :: Cond

wp'_Inh_StatementList :: Cond

data Syn_StatementList

Constructors

Syn_StatementList

Fields

copy_Syn_StatementList :: StatementList

length_Syn_StatementList :: Int

loopIterations_Syn_StatementList :: [Int]

wp_Syn_StatementList :: Cond

wrap_StatementList :: T_StatementList -> Inh_StatementList -> Syn_StatementList

sem_StatementList_Cons :: T_Statement -> T_StatementList -> T_StatementList

sem_StatementList_Nil :: T_StatementList